removed unused modules renamed functions2.py into functions.py
