--- src/synthetic_population_assignment.py
@@ -1,10 +1,0 @@
-
-
-class SyntheticPopulationAssignment:
-
-    def __init__(self, synthetic_population, assigned_data, modalities):
-
-        self._synthetic_population = synthetic_population
-        self._assigned_data = assigned_data
-        self._modalities = modalities
-
